requirements.txt
Django==2.1
djangorestframework
psycopg2
django-crispy-forms
django-gravatar2
gunicorn
whitenoise
django-filter
Pillow